Rename ModuleName to module in most cases
Summary:
Normally we store ModuleName in a variable called name. That's a bit weird, as the unique bit is the module-ness. Fix that everywhere, apart from in ModuleName and ModuleInfo, where name is indeed the unique bit.

Calling ModuleName a name is like calling R2-D2 (an R2 droid with designation D2) "R2" instead of "D2"...

Reviewed By: grievejia

Differential Revision: D69184480

fbshipit-source-id: a083f9bc22accca4839073eee272e30ca700948c