Make Typeshed a static global
Summary:
If we want the loader to load from TypeShed itself, we had better provide access to the same TypeShed. We have a single global one, so to avoid passing that around, actually make it global and cached.

If this turns out to be a bad idea, and we want to bundle two TypeShed, that won't be too hard later. But easier in the meantime to exclude it.

Reviewed By: grievejia

Differential Revision: D69184471

fbshipit-source-id: 898c1b2183b918d8f063962d174a48cbb92ef39f