Avoid resolving decorators twice during model query generation
Summary:
# Problem
Our model query DSL currently supports the `where = Decorator(fully_qualified_callee.equals(".."))` clause. This clause allows to match on the fully qualified name of the decorator applied to the function.
The implementation currently relies on resolving the call to the decorator during the model query execution.
This can be quite expensive, since this involves querying type information.

To handle decorators that are actually classes, we added a hack that resolves the expression again, adding the `__call__` call. This is leading to twice as many resolutions, which is expensive.

# Solution
Instead, we can use `new_targets` and `init_targets` provided by the first call resolution, and match on the `__init__` function. See test for an example.

Reviewed By: alexkassil

Differential Revision: D69189359

fbshipit-source-id: 9b92c6449732dabf4565b9b2bee9e62fdc1c019b