Allow loading modules from directories
Summary:
There are still many import-related errors in IG continuous jobs. They typically are of the form
```
from foo.bar import baz  # Error! Cannot find module foo.bar
```
But if we look at the filesystem structure, we can clearly see the existence of `foo/bar/baz.py`.

The problem here is that Pyre2 today still requires that all modules must have a corresponding path that's a regular *file* as its source. But in Python, certain modules may not have corresponding *file* on the filesystem -- they are called *namespace package* (see [PEP 420](https://peps.python.org/pep-0420)) and they corresponds to *directories* on the filesystem, not files. This kind of module is weird because
- It doesn't have `__file__` defined.
- There's no way to define `__all__` for it.
- In fact, there's no way to insert any kind of "initialization" logic to them, unlike `__init__.py`-backed modules (which are called *regular package*).

To be able to represent this kind of directory-backed modules, this diff introduces an additional module style called `Namespace`, and teach the loader to just use empty string as their contents instead of reading from filesystem (which would probably crash with some kind of "cannot read directory" error).

The rest of this stack will be aiming at changing the finder logic and have it construct & return this new kind of module. High-level description of the algorithm can be found in [this old post](https://fb.workplace.com/groups/657557118131740/permalink/1361339517753493/) I've written in the past.

Reviewed By: ndmitchell

Differential Revision: D69224141

fbshipit-source-id: d6994c4b961228cfd1bc7c7864687b7ae6d6e431