Search modules recursively in finder
Summary:
_recursion

Say if we have two search roots `[r0, r1]` and we want to find a module named `a.b` under them. Previously, Pyre2 simply attempts to check for the existence of these paths directly in order and return the first hit:
```
r0/a/b/__init__.pyi
r0/a/b/__init__.py
r0/a/b.pyi
r0/a/b.py
r1/a/b/__init__.pyi
r1/a/b/__init__.py
r1/a/b.pyi
r1/a/b.py
```
This is a good approximation for how the runtime searches for modules, but not quite the same. What [CPython would do at runtime](https://fb.workplace.com/groups/657557118131740/permalink/1361339517753493/) is to search in a "layer-by-layer" manner. For module `a.b`, it'll first try to locate where `a` is by checking these files:
```
r0/a/__init__.pyi
r0/a/__init__.py
r1/a/__init__.pyi
r1/a/__init__.py
```
If any one of those check hits, it'll use that as the "root" for the next phase of the search where it'll try to locate where `b` is under `a`. For example, let's say that `r0/a/__init__.py` exists, then we'll start from `r0/a/` and check the following paths for `a.b`:
```
r0/a/b/__init__.pyi
r0/a/b/__init__.py
r0/a/b.pyi
r0/a/b.py
```
Note that in this second phase of the search, `r1/a/` is not looked at all -- we've already "committed" to `r0/a/` when looking for `a`. The current algorithm used by Pyre2 finder wasn't able to emulate this subtlety today.

This diff attempts to replace what we have today with the "layer-by-layer" CPython algorithm described above. Not only is the new algorithm better aligned with CPython runtime, it's also very straightforward to extend it to support namespace packages (D69256720).

Reviewed By: ndmitchell

Differential Revision: D69224139

fbshipit-source-id: 3b508bb8a23a9cf2c2c79fbd055ff18441e1c3f5