Remove flow annot for loop phi keys
Summary:
We already scrubbed out everything except the annotation before, so I guess it's safe to keep scrubbing everything out?

The loop phi keys get merged after the loop, so this is part of the existing sketchiness about merging FlowStyle. My hunch is that there are bugs here (merging import styles, uninitialized flag, e.g.) but we only ran into issues with annots, so we only wrote code for those.

Reviewed By: yangdanny97

Differential Revision: D69272656

fbshipit-source-id: 1fec231d9eddfcc2b82f05fed8cc209808eef66c