Cache the resolution of decorators during model query execution
Summary:
# Problem
Our model query DSL currently supports the `where = Decorator(fully_qualified_callee.equals(".."))` clause. This clause allows to match on the fully qualified name of the decorator applied to the function.
The implementation currently relies on resolving the call to the decorator during the model query execution.
This can be quite expensive, since this involves querying type information.

Furthermore, we resolve the decorator call every time we evaluate a constraint against a target, which is very expensive.

# Solution
Let's "cache" the result of the resolution of the decorator call, so we perform this once when we evaluate a target against every model query. We can use `Lazy.t` for this purpose.

Reviewed By: tianhan0

Differential Revision: D69247516

fbshipit-source-id: 79b0aaf7ea38fd785d7cbce04d204d79957144ea