Move function solving to a new table
Summary:
This diff creates a new set of keys for functions, pointing to the solved type. This diff just introduces the necessary machinery, but otherwise leaves function solving alone.

For function bindings, we will calculate:
1. The type of the function with decorators applied
2. Metadata -- is this an overload?
3. Errors in cases where an overload implementation is not provided (via successor pointer stored on the binding)

(I think the metadata in 2 can also replace the Decoration stuff, but I will look into this later.)

We will still use Binding::Function to calculate the type of a function. In most cases, this will just forward to the new function binding. But overloaded functions will combine a chain of function bindings (via a predecessor pointer stored on the binding).

Reviewed By: ndmitchell

Differential Revision: D69451317

fbshipit-source-id: 5bb396cb91868ee4d693023bfb0f85f54bbc6af4