Add NoReturn binding check
Summary:
We decide on what binding to generate and whether to synthesize a None value or not by looking at the list of returns and the last statements of every branch of a function.

If we see any statements that could have a NoReturn type, we generate `ReturnExprWithReturn` containing those potentially never types and we solve the binding to generate the right subtyping checks, after we evaluate what the types we collected are, while taking into account all the other types in the binding.

For example if all branches have never types, then the type is never. If we have yields, then we pass the return portion of the generator as context as before.

Reviewed By: ndmitchell

Differential Revision: D69343257

fbshipit-source-id: b2bf0c70a8aec7cec39c9fe2cfff633ff73ab86f