Track duration of expression analysis with the profiler
Summary:
We currently have a profiler that can track the time spent analyzing each statement, in the forward and backward analysis. It is enabled by adding a call to `pyre_dump_perf()` in the function to profile.

However, there are cases where analyzing at the statement level is still too imprecise, and we would want to know exactly which expression was slow to analyze.

This diff implements a profiler for expression. Each call to `analyze_expression` will be logged, alongside the time spent into it.
We only measure the time spent on the expression itself, and exclude nested calls.

Reviewed By: alexkassil

Differential Revision: D69466835

fbshipit-source-id: 87018918a43c1f713a40c02777311a296e429696