Disallow redirecting to decorated functions for the recognized calls
Summary:
If a `CallCallees.t` is merged from two `CallCallees.t` where one is a
recognized call and the other one is not, it is unclear whether it is better to
redirect the merged one.
Since it is probably rare for a `CallCallees.t` to be a recognized call, it
might be better to redirect the merged one, so that the majority of the calls
are properly redirected, at the expense of also redirecting the few ones that
should not be redirected.

Reviewed By: arthaud

Differential Revision: D69127908

fbshipit-source-id: 70108c7e59299a6afa2042f63fe39bf71e5bee56