Add heuristic to ignore dunder calls in higher order parameters
Summary:
# Problem
We currently have a heuristic to prevent false negatives with functions being passed as arguments of other functions. We assume the function-as-a-parameter is called at the call site, i.e:
```
def foo(): return source()
def bar(f): ... # no code
def baz():
  bar(foo) # assume foo() is called
```
Without this heuristic, we could have false negatives since we don't see the code of `bar`. Even if we did, we don't track function-as-values interprocedurally (note: we are actually working on this).

The problem is that the heuristic also treats instances of callable classes as functions, i.e:
```
class A:
  def __call__(self): return source()

def foo(x: A): pass
def bar():
  x = A()
  foo(x)
```
At the call to `foo(x)`, we will introduce an artificial call to `A.__call__`. This can lead to false positives, as well as performance problems, adding a lot of invalid edges in the call graph.

# Solution
We still need the heuristic in cases where the instance is passed to an unknown functions.
However, when the instance is passed to a function with a body and properly annotated (such as the example above), then we would be able to see if `__call__` was called. Therefore, it is not necessary to add the artificial call to `__call__`.

This diff implements this check for trivial cases.

Reviewed By: tianhan0

Differential Revision: D69538135

fbshipit-source-id: b1fbab23ef893c5a3a152b10860ab406d37d991e