Update BindingClassField comments
Summary:
`BindingClassField` includes fields and methods which was a bit confusing to me. But I found the full list of terminology and it is located in class.rs.

I would hesitate to link to this here in case the location changes (or should we do so?), but wanted to clarify that in BindingClassField we are talking about all attributes such as methods as well  incase someone doesn't see the comment there.

Reviewed By: rchen152

Differential Revision: D69622922

fbshipit-source-id: 92ae7ecb4c8145c6334ee8bf582bac0830b9dd2b