Move `get_import` from `answers.rs` to `attr.rs`
Summary:
justmovingthingsaround

`get_import()` is defined in `answers.rs` but its only caller is in `attr.rs`. The only reason why it's placed in `answers.rs` is because the data it tries to read from -- the `exports` map -- is marked as private in `answers.rs`.

I think marking the `exports` map public and change `get_import()` into a private function in `attr.rs` makes more sense, as it'll eliminate lots of boilerplates in subsequent diffs just to allow `attr.rs` get partial access into the info that's available in `exports`.

Reviewed By: ndmitchell

Differential Revision: D69564697

fbshipit-source-id: cbf97f5b8745885dd49a9163f32dd74df1284998