Error when accessing nonexistent attributes on modules
Summary:
the_limit_does_not_exist

In D69564695 I mentioned that we currently don't perform any existence check on submodule accesses. i.e. if `foo` is a package but there's no module that corresponds to `foo.bar`, we don't emit any error which is clearly not ideal. This diff attempts to add minimal amount of checks here to make the overall behavior a bit more reasonable for now.

Reviewed By: ndmitchell

Differential Revision: D69564701

fbshipit-source-id: bb6e4bbbf709296539a47ef82506a7b8d6025401