Always schedule jobs on a worker, even when number_of_chunks=1
Summary:
# Problem
We recently found a performance bottleneck in Pysa due to copying/marshal-ing large closures to each workers.
Most notably, the `pyre_api` handle would take a few dozen seconds to copy from the main process to a worker. It seems like it accumulates quite a lof of data during pysa's pre-analysis steps. See D69659719 for more context.

One reason for `pyre_api` becoming large was that some `map_reduce` operations would actually be executed by the main process, when `number_of_chunks = 1`.
One concrete example is the parsing of `.pysa` files, when the number of files to parse is low. Parsing `.pysa` files actually leads to quite a lot of calls to pyre APIs, which would lead to caching, stored in `pyre_api`. That cache is preserved for the whole analysis, and copy-ed to each worker.

# Solution
Let's always use a worker in a `map_reduce` call, even when `number_of_chunks=1`. The only exception is when the scheduler is not a `ParallelScheduler`.

Reviewed By: tianhan0

Differential Revision: D69659720

fbshipit-source-id: 62d5b9201e3d2fe712ea7a279bc2586aa4913725