Use map-reduce for all operations requiring the pyre API handle
Summary:
# Problem
We recently found a performance bottleneck in Pysa due to copying/marshal-ing large closures to each workers.
Most notably, the `pyre_api` handle seems to take a few seconds to copy from the main process to a worker.
It accumulates quite a lot of data (probably caching?) during steps that are executed on the main process.

# Solution
To avoid this problem, we should almost-always perform operations that require `pyre_api` within a map-reduce, since workers are killed after the map-reduce. Let's do that in this diff.

Reviewed By: tianhan0

Differential Revision: D69659719

fbshipit-source-id: 1ac5cf6a1630f611f922fdb72015410d571a6493