Make it clear where the drop happens
Summary:
I'm sure Rust specifies a drop order, but I'm not sure what it is. I think it's actually the opposite of what we want. Do make it clear we drop without the lock held.

In practice we expect the Ast to be grabbed by the computation, but better to be clear on these things.

Reviewed By: stroxler

Differential Revision: D69741627

fbshipit-source-id: 0c39a0f30d2dc547a5b1074782a1ea9a47e0f0d4