Use protocol subtyping to implement iteration
Summary:
Now that we have protocol support, we can use it to simplify the `iterate` implementation. Before, we synthesized calls to the `__iter__` and `__next__` methods. Now, we create a var and solve it using is_subset_eq, like other unwrap operations.

I also took the opportunity to eliminate the catch-all match case, which unconditionally failed. Instead, we now check everything against the iterator protocol.

We still have the fallback for `__getitem__` iteration. It actually might be nice to have a protocol definition for this in typeshed, but the synthesized call is ok for now.

Reviewed By: ndmitchell

Differential Revision: D69760235

fbshipit-source-id: e7f6b1b5d48bfe071311bea40a085317db61054b