Store initial value of assignments in FlowStyle
Summary:
I want to store `Expr`s that class fields are initialized to in `BindingClassFieldInitialization` so I can get at `dataclasses.field(...)` expressions. `BindingClassFieldInitialization` is computed from `FlowStyle`, so the first step is to store the expression in the latter.

Context: I'm working on support for `dataclasses.field`, which lets you specify properties of dataclass fields as keyword arguments to a magic `field` function:
```
dataclass
class C:
    x: int = field(init=False)  # x will not appear in C.__init__
```
Two things make this tricky:
* We can't get the keyword arguments from the resolved type of the `field` call, since it just resolves to whatever the default value of the field is (e.g., `field(default=0)` ends up with type `int`, with no information about the fact that the `field` function was called).
* We can't special-case `dataclasses.field` in a way that can't be extended to other functions, because we eventually need to support `dataclass_transform`.

With these requirements, the best idea I could come up with is to change `ClassFieldInitialization` to compute and store the dataclass field properties.

Reviewed By: ndmitchell

Differential Revision: D69678234

fbshipit-source-id: bd3800e2bc5031e82c1c7ca13bdd46b9bd1525b0