allow star-unpacking Type::Quantified and Type::TypeVarTuple
Summary:
This diff changes the representation of Tuple::Unpacked to have the middle represent the type being unpacked, as opposed to the type of each middle element. This allows us to have unpacked tuples where the middle is a TypeVarTuple.

This diff also rewrites the subset logic to account for this, and also changes the `Tuple::unpacked` factory to return just the middle type if both the prefix and suffix are empty.

Later diffs in the stack will work on matching type arguments to type params with TypeVarTuple.

Reviewed By: ndmitchell

Differential Revision: D69788505

fbshipit-source-id: 35fb7026a2afb92fada474c11a2403cedc6f9785