Add failing test showing bad overload selection behavior
Summary:
This stack:

When we analyze a call, we also infer the types of the arguments. For overloaded calls, we try to find an overload where the call succeeds by checking if there are any errors.

This behavior is too strict. An argument expression might result in an error, but still evaluate to a type that is compatible with the function signature. These errors should not affect overload selection.

This diff:

Adds a test :)

Reviewed By: ndmitchell, rchen152

Differential Revision: D69901522

fbshipit-source-id: aefd02cc36dec783a83d14de68354b39827945fd