Align test stubs with real stubs for `Any`
Summary:
We wasted some time today tracking down an issue that was caused by a code change that
worked fine for our stubs, which say `Any` is a class, but don't work against trunk
typeshed which says `Any = object()`; currently Pyre2 makes it much easier to identify
classes "by identity" (e.g. in `stdlib.rs` whereas normal values cannot be identified
easily.

In reality, `Any` *is* a class, which makes this seem like a bug in the stubs, and we're
going to look into patching upstream, but in the meantime we really ought to use the
same definition for tests that typeshed has, and chart a path forward that works for both.

Reviewed By: rchen152

Differential Revision: D69954969

fbshipit-source-id: 7cfef31c50f06f44f4ece9d20a092837f9665633