handle getitem call for unpacked tuple
Summary:
`as_attribute_base` is a bit tricky, since if the middle is an unpacked TypeVarTuple we want an AttributeBase::Quantified, but we can't account for both the concrete prefix/suffix and the middle at the same time since we can only return one AttributeBase. For now, we fall back to the ole-reliable `tuple[Any, ...]`.

I think if we could return multiple AttributeBases and plumb that behavior through LookupResult and union the final result it could work, but that's a large change.

Reviewed By: stroxler

Differential Revision: D69937151

fbshipit-source-id: ad49614c65430e62af764b2688a39b707c10885d