Setup structs to trace attribute definitions
Summary:
In this stack, we will implement basic support for go-to-definition on attributes.

The definition might come from another module, so we need to track both the module information and the `TextRange` within that module. Therefore, I created `TextRangeWithModuleInfo` struct for this purpose.

We start by setting up the place to store all the definition locations. Similar to the traces we added for types in D69924937, we store the traces for definitions in the same place, together with types in a `Traces` struct.

Reviewed By: stroxler

Differential Revision: D69948907

fbshipit-source-id: f970ff34da5d3cde1d6e8cc37b5494eb7f972f14