Split ClassField logic into its own module
Summary:
The `classdef.rs` module is still very big even after splits we've done recently, and
after looking at it closely this is largely because it has two mostly orthogonal
responsiblities: dealing with the intricacies of `ClassField`, which is already complex
and will likely become more so, and dealing with the wholistic "glue" of combinging
`ClassField`, `ClassMetadata`, and `Class` to actually analyze things like attribute
access.

We can split it almost exactly in half by cutting along this access: everything local
to a single class field is now handled in the `class_field.rs` module, including the
datatype definitions and many of the simple helper function.

The downstream handling of class "members", class attributes, instantiation, etc are
still in `classdef.rs` because that's the point at which we start having to combine
class field information with class metadata.

Reviewed By: yangdanny97

Differential Revision: D70042287

fbshipit-source-id: ad1c20058dc6d7851ae0354e9cd95c1d67112045