Use `__classproperty__` for Pyre1 unit tests
Summary:
In D69931339 I removed support for `pyre_extensions.classproperty`,
which doesn't even exist; this is important because it's confusing
to have references to nonexistant code.

But we *do* need to have some class property available for unit
tests, and since there exists no class property in the standard
library it makes sense to use a placeholder name just so we can
test the functionality.

I decided to go with unqualified `__property__` for this, which is
consistent with the mechanism we've used for similar use cases
in the past, such as `__dataclass_transform__` when that PEP was
incomplete and there was not yet a cannonical definition to use.

There are lots of trivial line number changes in one of the Pysa integration
tests; that's just because I had to add a `pyre-ignore`, because Pyre
understands the decorator at signature time but also (correctly) complains
that this name is undefined at type-check time

Reviewed By: yangdanny97

Differential Revision: D70040415

fbshipit-source-id: 52ff8c87df2c6f01d352fa8a7226cd2a842f8f61