Add failing tests for contextual typing from generic function return types
Summary:
For generic function calls, we can use a return type hint to help instantiate the type arguments. This is only really useful if the other arguments to the function do not provide any information.

I am not planning to work on this in the immediate future, so leaving the tests behind to define the future work.

Reviewed By: ndmitchell

Differential Revision: D70057714

fbshipit-source-id: 768ba035a8a2a12c58525cd1f8e394f7e95a8c51