Use helpers to hide ClassField internals from typed dict code
Summary:
This diff continues my work toward making ClassField internals private (which makes it
much easier to iterate on the representation) by making a clear API for various use cases
of special-case class constructs that need non-standard analysis.

Here, we create a `TypedDictFieldInfo` abstraction and a helper method. Note that `TypedDictFieldInfo`
is different from `TypedDictField` in that (a) it is private to the `class` module and sub-modules,
(b) it is field-level only in analysis scope, like `ClassField` - so for example `required` is an option
because we don't analyze the class-level default in the same layer as the annotation, and (c) it has
an extra bit of information about initialization that's needed to synthesize the correct `__init__`
constructor.

NOTE: because this is a refactor diff in a refactor stack, I'm preserving behavior, but the
bug identified in <FILLME> was caused by inappropriate sharing of the `as_param` helper between typed
dict and dataclasses; this commit decouples the two so I'll stack a fix in the next commit.

Reviewed By: ndmitchell

Differential Revision: D70057866

fbshipit-source-id: 05fcfc914fbe08c2006037c3316b1ff986708dad